launch repository 분리
