+++ app/src/main/java/com/jin/honey/MainActivity.kt
@@ -9,21 +9,22 @@ import androidx.lifecycle.lifecycleScope
 import androidx.room.Room
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
-import com.jin.database.AppDatabase
 import com.jin.data.address.AddressRepositoryImpl
 import com.jin.data.addressimpl.AddressDataSourceImpl
 import com.jin.data.cart.CartRepositoryImpl
-import com.jin.datastore.PreferencesRepositoryImpl
+import com.jin.data.chat.ChatRepositoryImpl
 import com.jin.data.firestoreimpl.FireStoreDataSourceImpl
 import com.jin.data.food.FoodRepositoryImpl
-import com.jin.honey.feature.navigation.RootNavigation
-import com.jin.network.kakao.KakaoMapApiClient
-import com.jin.network.openai.OpenAiApiClient
-import com.jin.data.chat.ChatRepositoryImpl
 import com.jin.data.openaiimpl.OpenAiDataSourceImpl
 import com.jin.data.order.OrderRepositoryImpl
 import com.jin.data.recipe.RecipeRepositoryImpl
 import com.jin.data.review.ReviewRepositoryImpl
+import com.jin.database.AppDatabase
+import com.jin.datastore.PreferencesRepositoryImpl
+import com.jin.datastore.launch.LaunchRepositoryImpl
+import com.jin.honey.feature.navigation.RootNavigation
+import com.jin.network.kakao.KakaoMapApiClient
+import com.jin.network.openai.OpenAiApiClient
 import com.jin.ui.theme.HoneyTheme
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
@@ -56,6 +57,7 @@ class MainActivity : ComponentActivity() {
                         db.foodTrackingDataSource(),
                         FireStoreDataSourceImpl(firestore)
                     ),
+                    launchRepository = LaunchRepositoryImpl(this),
                     preferencesRepository = PreferencesRepositoryImpl(this),
                     cartRepository = CartRepositoryImpl(db.cartTrackingDataSource()),
                     addressRepository = AddressRepositoryImpl(
