+++ app/src/main/java/com/jin/honey/feature/onboarding/ui/OnboardingViewModel.kt
@@ -2,6 +2,7 @@ package com.jin.honey.feature.onboarding.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.jin.domain.launch.LaunchRepository
 import com.jin.domain.repositories.PreferencesRepository
 import com.jin.domain.usecase.SyncAllMenuUseCase
 import com.jin.domain.usecase.SyncRecipesUseCase
@@ -9,7 +10,7 @@ import com.jin.domain.usecase.SyncReviewsUseCase
 import kotlinx.coroutines.launch
 
 class OnboardingViewModel(
-    private val repository: PreferencesRepository,
+    private val repository: LaunchRepository,
     private val syncAllMenuUseCase: SyncAllMenuUseCase,
     private val syncReviewsUseCase: SyncReviewsUseCase,
     private val syncRecipesUseCase: SyncRecipesUseCase
