+++ domain/src/main/java/com/jin/domain/launch/LaunchRepository.kt
@@ -1,0 +1,6 @@
+package com.jin.domain.launch
+
+interface LaunchRepository {
+    suspend fun completeFirstLaunch()
+    suspend fun isFirstLaunch(): Boolean
+}
