+++ domain/src/main/java/com/jin/domain/repositories/PreferencesRepository.kt
@@ -3,12 +3,11 @@ package com.jin.domain.repositories
 import kotlinx.coroutines.flow.Flow
 
 interface PreferencesRepository {
-    suspend fun completeFirstLaunch()
-    suspend fun isFirstLaunch(): Boolean
     fun flowSearchKeywords(): Flow<List<String>>
     suspend fun saveSearchKeyword(menuName: String)
     suspend fun deleteSearchKeyword(menuName: String)
     suspend fun clearSearchKeyword()
+
     suspend fun insertOrUpdateFavoriteMenu(menuName: String)
     fun flowFavoriteMenus(): Flow<List<String>>
     suspend fun insertRecentlyMenu(menuName: String)
